create src and test folders and move php scripts there
