+++ src/init-test.php
@@ -1,0 +1,6 @@
+<?php
+
+$upOneDir = dirname(__DIR__ );
+require_once $upOneDir . "/vendor/autoload.php";
+
+require_once "src/Task.php";
